[fix] restructure the file structure in front end src
